[AGP DSL] Fix lint block in groovy
This regressed with Ia8b37b693e6c841eb7be67c0aee5c6acd8b961d1.

The AGP DSL decorator automatically generates an Action taking method
that triggers Gradle's instantiator to generate a groovy closure so
users get 'this' style rather than 'it'.

However, for CommonExtension we use Kotlin delegation from
the legacy BaseExtension hierarchy of classes, so that
generated method isn't on the extension as accessible by the user.

Adding it to InternalCommonExtension, makes the Kotlin compiler generate
a delegate method, and causes Gradle to generate the appropriate method,
so it all works.

Bug: 140406102
Test: Updated LintBaselineTest to use the new DSL
Change-Id: Ia8b37b693e6c841eb7be67c0aee5c6acd8b961d1