Move maven repo building to build time
Before this change maven_repo would generate a list
of files to be included in a maven repo together with
metadata for them (classifiers) at build time. At
runtime (tests) this file was taken to resolve a maven
repo layout, and then create symlinks for it.

This change moves the resolution phase to build time
and the generated file is already in repository layout.

This now only allows for the new incoming maven rule set
to work but simplifies several things:

. RepoLinker modules are gone, as it is now a simple utility
  function in test utils
. Tests do not have to depend on maven libraries to know how
  to layout a repository.
. The manifest files are simple _zipper argument lists
. The maven_repo_zip rule became an optional output of the rule

Test: existing tests + updated RepoLikerTest
Bug: 190268212
Change-Id: I45f264b86c7371fa912fed19faff6ad52085ae09