Merge GlobalLibraryMap into VariantDependencies
With upcoming caching on a per-project basis for
the sync model, we cannot have a global model that's
queried on a random project.
Because of this, this change moves the library map
to be computed and returned on a per-project basis.

It's still advantageous to keep the GraphItem as
light-weight as possible, and remove redundancy
across the graphs for the main, test, androidTest and
fixture artifacts.

This change also renames GlobalLibraryBuildService to
GlobalSyncService since it's not tied to the Library
instances anymore but still provides services to sync.

Finally, tweak and improve the test fixtures for
querying model and comparing them.

Bug: N/A
Test: existing tests. Updated the snapshots.
Change-Id: I2c5d9a53670e90e92e5122fbc343ccb598fb488f