Ensure Kotlin stdlib paths are absolute from unit tests
When lint's unit tests are running and the unit tests refer to
Kotlin, it automatically adds the Kotlin stdlib into the class
path. It does this by pulling the kotlin libraries from the
environment. In some cases, those paths may be relative, so
make them absolute first (since they're required to be absolute
when reaching the UAST environment setup).

Test: Existing
Bug: None
Change-Id: Ie84ee5eda64ab9e494f45f1f878e0759f71bc3c3