Fix LintJar in v2 model.
The information is gathered by querying for
the lint jars via the ArtifactCollection API
rather than via the DependencyResult API where
no artifact data is possible.
Using the same Variant Key we can inject the
lint jar path into the custom ResolvedArtifact
class.

This replaced the hardcoded path for lint jars
inside of external AARs with a side effects
that the value is now non null only if there
is actually a lint jar inside an AAR.

For sub-projects, the value is present as well
but because Library modules always publish a
lint jar even if nothing is attached to the
lintPublish Configuration, the file may not
actually exist.

Fixes: 191006100
Test: new integ tests
Change-Id: Ib17b39901709f37bf046df62f0ecad39813a0ba4