[Lint] Add if/else detection in ConstantEvaluator
Previously, if ConstantEvaluator encountered an assignment in an
if/else block, it will always evaluate the else branch, even if
that assignment is not always executed. Furthermore, if an
assignment cannot be determined statically by the evaluator, it will
no longer resort to the initializer value, since it's possible that
that value is always replaced in an if/else block and thus unreachable.

(Please see the unit tests for such scenarios.)

Bug: 197179112
Test: In commit
Change-Id: I4f6743376febe087a6e30e674b6eca971ea72f39