Fix bug where extra ABIs were built
If variant debug1 targets x86 and debug2 targets x86_64 but are otherwise compatible configurations then the prior task dependency graph would be:

debug1 and debug2 depend on
  buildCMakeDebug (builds x86 and x86_64) depends on
    configureCMakeDebug (configures x86 and x86_64)

So that when the user builds debug1 they also get an unwanted configure and build for x86_64.

After this fix, the task dependency graph is:

debug1 depends on
  buildCMakeDebug[x86] depends on
    configureCMakeDebug[x86]
debug1 depends on
  buildCMakeDebug[x86_64] depends on
    configureCMakeDebug[x86_64]

So that building debug1 results in only x86 being configured and built.

A side benefit is that the user will have reliable per-ABI task names for configure and build.

Note that, for build, there is still an extra dimension for build targets. If variant debug3 build target-1 and target-2 and variant debug4 builds target-2 and target-3 then the resulting tasks are:

 debug3 depends on
   buildCMakeDebug[ABI][target-1,target-2]
      configureCMakeDebug[ABI]
 debug4 depends on
   buildCMakeDebug[ABI][target-2,target-3]
      configureCMakeDebug[ABI]

So there's not the problem of extra targets being built but there is the possible problem that there are two tasks that build target-2. I decided against breaking out the target dimensions into individual tasks (at least for now) because doing so would defeat Ninja's build parallelization. I don't see a way that this could be a problem since build has natural up-to-date checking. I'm just making a note for future reference.

Bug: 195121515
Test: Added test and adjusted existing
Change-Id: I4a95a6d927d88424aa7943f6414a48b849c3e911